requirements.txt
APScheduler==3.6.3
cachetools==4.2.2
certifi==2023.5.7
flake8==6.0.0
mccabe==0.7.0
pycodestyle==2.10.0
pyflakes==3.0.1
python-dotenv==1.0.0
python-telegram-bot==13.7
pytz==2023.3
six==1.16.0
tornado==6.3.2
tzlocal==5.0.1
